fix(gtk4): move __version__ to leaf module to break remaining cyclic imports
CI's last failing check (CodeQL py/cyclic-import at clipman/window.py:20)
came from `from clipman import _, __version__, updates` re-entering the
package root. Same family of cycle the previous fix broke for
edge_states.py, just on the other submodules that still touched the
package root directly.

Restructure so no submodule needs `from clipman import …`:

- New clipman/_version.py — single source of truth for __version__.
  clipman/__init__.py re-exports it so the public clipman.__version__
  API is unchanged (CLI --version, tests, pyproject metadata all
  resolve the same).
- clipman/window.py, clipman/preferences.py, clipman/snippets_dialog.py:
  swap `from clipman import _` for `from gettext import gettext as _`
  (clipman/__init__.py has already called textdomain("clipman") by
  the time submodules load, so the translation domain is correct).
  Swap `from clipman import __version__` for `from clipman._version
  import __version__`.
- clipman/updates.py: same _version swap.
- scripts/bump-version.sh: patches clipman/_version.py instead of
  clipman/__init__.py (the literal moved files).

ruff clean, py_compile clean, `import clipman; clipman.__version__`
still returns "1.0.6".
