+++ clipman/__init__.py
@@ -1,9 +1,10 @@
 import gettext
 import os
 
-# Source of truth for the running daemon version.
-# scripts/bump-version.sh keeps this in sync with pyproject.toml.
-__version__ = "1.0.6"
+# Public version constant — re-exported from the leaf ``_version`` module
+# so submodules can import ``__version__`` without re-entering the
+# package root (CodeQL py/cyclic-import).
+from clipman._version import __version__
 
 LOCALE_DIR = os.path.join(
     os.path.dirname(os.path.dirname(os.path.abspath(__file__))), "locale"
@@ -11,3 +12,5 @@ LOCALE_DIR = os.path.join(
 gettext.bindtextdomain("clipman", LOCALE_DIR)
 gettext.textdomain("clipman")
 _ = gettext.gettext
+
+__all__ = ["__version__", "_", "LOCALE_DIR"]
